made changes to dashboard + atomic bomb

+++ FRC_2024_Crescendo/src/main/java/frc/robot/Constants.java
@@ -135,10 +135,10 @@ public final class Constants {
     public static final int kLimitSwitchControllerPort = 9;
 
     // Change later after arm built -------------------------------------
-    public static final double kArmPickup = 24;
-    public static final double kArmAmp = 16;
-    public static final double kArmUp = 12;
-    public static final double kArmShooter = 2;
+    public static final int kArmPickup = 24;
+    public static final int kArmAmp = 16;
+    public static final int kArmUp = 12;
+    public static final int kArmShooter = 2;
 
     public static final double kVoltageMultiplier = 1.5;
   }
